♻️ extract location.specHelper
