+++ packages/core/test/index.ts
@@ -3,6 +3,7 @@ export * from './buildEnv'
 export * from './capturedExceptions'
 export * from './collectAsyncCalls'
 export * from './createNewEvent'
+export * from './location.specHelper'
 export * from './requests.specHelper'
 export * from './stubReportApis'
 export * from './stubZoneJs'
